requirements.txt
fastapi
uvicorn[standard]
motor
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
pydantic-settings
python-multipart